fixing type variables

+++ 2-selection_sort.c
@@ -6,7 +6,7 @@
  */
 void selection_sort(int *array, size_t size)
 {
-	unsigned int i, temp, j;
+	size_t i, temp, j;
 	int state;
 
 	if (array == NULL || size <= 1)
